backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.7.0
python-multipart>=0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database & Embeddings
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.5.23

# LLM Integration
openai>=1.3.0
tiktoken>=0.5.0
ollama>=0.4.0

# Data Processing
GitPython>=3.1.0
python-docx>=1.1.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiofiles>=23.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.27.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
